create unifyfs-stage and unify.c calls it
Creates unifyfs-stage infrastructure and binary.  That's used
to stage files in and out at the startup and termination of
unifyfs, respectively.

The unifyfs (sourced in unifyfs.c) calls that binary through
command line flags --stage-in and --stage-out.

Because unifyfs-stage has ssl-based tools for verification that
files were transferred correctly, now have a dependency on
an SSL package.