remove client write extents when truncating
Truncation is a synchronization point, so unsynced extents
are passed to the server before performing the truncate.
However it is inefficient to pass extents that we are only
going to delete as part of the truncate. This adds support
for deleting/updating client extents that are affected by
the truncation. A new seg_tree_remove() method contains
all the logic to match and update/remove extents that
overlap a given range, and truncate is treated as a
remove from [trunc_offset, ULONG_MAX].